requirements.txt
fastai==2.8.7
plum-dispatch==2.8.0
fasttransform==0.0.2
cloudpickle==3.1.2
ipython==9.12.0
